refactor: simplify svelte definition
