--- packages/svelte-flicking/src/FlickingPanel.js
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2015 NAVER Corp.
- * egjs projects are licensed under the MIT license
- */
-import FlickingPanel from "./flicking-panel.svelte";
-
-export default FlickingPanel;
