+++ packages/svelte-flicking/src/flicking-panel.svelte.d.ts
