+++ packages/svelte-flicking/src/index.ts
@@ -3,7 +3,7 @@
  * egjs projects are licensed under the MIT license
  */
 import Flicking from "./Flicking";
-import FlickingPanel from "./FlickingPanel";
+import FlickingPanel from "./flicking-panel.svelte";
 
 export default Flicking;
 export {
