+++ packages/svelte-flicking/src/index.umd.ts
@@ -3,7 +3,7 @@
  * egjs projects are licensed under the MIT license
  */
 import Flicking from "./Flicking";
-import FlickingPanel from "./FlickingPanel";
+import FlickingPanel from "./flicking-panel.svelte";
 
 (Flicking as any).FlickingPanel = FlickingPanel;
